migrate from uint32 to int
instead of just converting the ints to uint32 and back, we could use
ints directly

also create a new helpers file which is used by main.go

+++ stellar/database/originators.go
@@ -33,7 +33,7 @@ func (originator *ContractEntity) OriginContract(panelSize string, totalValue in
 	if len(allOrders) == 0 {
 		pc.O.Index = 1
 	} else {
-		pc.O.Index = uint32(len(allOrders) + 1)
+		pc.O.Index = len(allOrders) + 1
 	}
 
 	pc.O.PanelSize = panelSize
